average rating is a float type

+++ src/schema/movie.ts
@@ -5,7 +5,7 @@ const typeDef = `
         duration: String!
         releaseDate: String!
         actors: String!
-        averageRating: Int
+        averageRating: Float
     }
 
     type DeleteResponse {
